new main with helpers
